typecast float to int in print_msg

+++ 0x03-pathfinding/100-a_star_graph.c
@@ -118,7 +118,7 @@ static void print_msg(a_node_t *node, const vertex_t *start)
 		printf("Checking %s, distance from %s is %d\n",
 				(char *)node->vertex->content,
 				(char *)start->content,
-				node->distance);
+				(int)node->distance);
 	}
 }
